This commit brakes several tests:
perceived?/possible? problem:
A discrete_model_run that reproduces the solution uses net_Us (or
equivalently the solution xs) and can not reproduce the gross_Us that
might also be available via the data.
On the other hand a discrete model run that wants to use only the
gross_Us needs to reproduce the net_Us to faithfully reproduce the
solution, but needs knowledge of the state transition operator within
the time step (The B_s are only Phis from the beginning of the time step
to its end) to compute the net_Us and at least has to make an assumption
(of constancy) about the infinitesimal B_pwc on the interval.

contradictory claim:
If we use gross_Us to generate a discrete_model_run via
reconstruct_from_fluxes we can reproduce the solution of the
pwc_model_run that had been used to fake the data via _fake_discretized_output.
(although _fake_discretized_output uses gross_Us whereas
DiscreteModelRun.from_PWCModelRun used net_Us)